clean: move get_shape to asteroid.utils.torch_utils
